backend/requirements.txt
fastapi
uvicorn
google-generativeai>=0.8.3
python-dotenv